♻️ (#196) 키 업데이트 및 폴더 이동
