--- src/features/task/api/taskQueryKeys.ts
@@ -1,20 +1,0 @@
-export const TASK_QUERY_KEYS = {
-  root: ['tasks'] as const,
-
-  project: (projectId: string, status: string) =>
-    [...TASK_QUERY_KEYS.root, 'project', projectId, status] as const,
-
-  meStatus: (status: string) => [...TASK_QUERY_KEYS.root, 'me', status] as const,
-
-  member: (projectId: string, memberId: string) =>
-    [...TASK_QUERY_KEYS.root, 'project', projectId, 'member', memberId] as const,
-
-  detail: (projectId: string, taskId: string) =>
-    [...TASK_QUERY_KEYS.root, 'detail', projectId, taskId] as const,
-
-  projectCountStatus: (projectId: string) =>
-    [...TASK_QUERY_KEYS.root, 'project', projectId, 'countStatus'] as const,
-
-  projectCountMember: (projectId: string) =>
-    [...TASK_QUERY_KEYS.root, 'project', projectId, 'countMember'] as const,
-};
