+++ src/features/task/hooks/useTaskDetailQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 import { taskApi } from '@/features/task/api/taskApi';
-import { TASK_QUERY_KEYS } from '@/features/task/api/taskQueryKeys';
+import { TASK_QUERY_KEYS } from '@/features/task/constants/taskQueryKeys';
 
 // 할 일 상세 조회
 export const useTaskDetailQuery = (projectId: string, taskId: string) => {
